Move MysqlPool to backend
