+++ core/src/db/meta/backend/MySQLConnectionPool.cpp
@@ -9,9 +9,11 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
 
-#include "db/meta/MySQLConnectionPool.h"
-#include <fiu-local.h>
+#include "db/meta/backend/MySQLConnectionPool.h"
+
 #include <thread>
+
+#include <fiu-local.h>
 
 namespace milvus::engine::meta {
